+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/resourcePolicies/ExecutionClusterInfo.scala
@@ -1,9 +1,3 @@
 package edu.uci.ics.amber.engine.architecture.scheduling.resourcePolicies
 
-object ExecutionClusterInfo {
-  def apply(): ExecutionClusterInfo = {
-    ExecutionClusterInfo()
-  }
-}
-
-case class ExecutionClusterInfo() {}
+class ExecutionClusterInfo() {}
