+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/resourcePolicies/ResourceAllocation.scala
@@ -1,40 +1,35 @@
 package edu.uci.ics.amber.engine.architecture.scheduling.resourcePolicies
 
-import edu.uci.ics.amber.engine.architecture.scheduling.Region
+import edu.uci.ics.amber.engine.architecture.scheduling.{Region, RegionConfig, WorkerConfig}
 import edu.uci.ics.amber.engine.common.AmberConfig
+import edu.uci.ics.amber.engine.common.virtualidentity.PhysicalOpIdentity
 import edu.uci.ics.texera.workflow.common.workflow.PhysicalPlan
 
-import scala.collection.mutable.ListBuffer
+class ResourceAllocator(physicalPlan: PhysicalPlan, executionClusterInfo: ExecutionClusterInfo) {
 
-object ResourceAllocation {
-  def apply(region: Region): ResourceAllocation = {
-    ResourceAllocation(region)
-  }
-}
-
-case class ResourceAllocation(region: Region) {
-
-  private def physicalOpIds = region.physicalOpIds.filter(_.logicalOpId.id.contains("PythonUDF"))
-
-  def allocation(
-      physicalPlan: PhysicalPlan,
-      executionClusterInfo: ExecutionClusterInfo
-  ): (List[Int], Double) = {
-    val workers = ListBuffer[Int]()
-    physicalOpIds.foreach { opId =>
-      val operator = physicalPlan.getOperator(opId)
-
-      if (operator.parallelizable) {
-        operator.suggestedWorkerNum match {
-          case Some(num) => workers += num
-          case None =>
-            workers += AmberConfig.numWorkerPerOperatorByDefault // Set to default value if not defined
+  def allocate(
+      region: Region
+  ): (Region, Double) = {
+    val config = RegionConfig(
+      region.getEffectiveOperators
+        .map(physicalOpId => physicalPlan.getOperator(physicalOpId))
+        .map { physicalOp =>
+          {
+            val workerCount =
+              if (physicalOp.parallelizable) {
+                if (physicalOp.suggestedWorkerNum.isDefined) {
+                  physicalOp.suggestedWorkerNum.get
+                } else {
+                  AmberConfig.numWorkerPerOperatorByDefault
+                }
+              } else {
+                1
+              }
+            physicalOp.id -> (0 until workerCount).map(_ => WorkerConfig()).toList
+          }
         }
-      } else {
-        // Set to 1 if not parallelizable
-        workers += 1
-      }
-    }
-    (workers.toList, 0)
+        .toMap
+    )
+    (region.copy(config = Some(config)), 0)
   }
 }
